rename mesh-identity and mesh-trust modules
