+++ packages/kuma-gui/src/app/mesh-identities/index.ts
@@ -8,7 +8,7 @@ type ResourcesSources = ReturnType<typeof sources>
 
 export const services = (app: Record<string, Token>): ServiceDefinition[] => {
   return [
-    [token<ResourcesSources>('mesh-trust.sources'), {
+    [token<ResourcesSources>('mesh-identities.sources'), {
       service: sources,
       arguments: [
         app.api,
