+++ packages/kuma-gui/src/app/meshes/index.ts
@@ -12,8 +12,8 @@ import { services as dataplanes } from '@/app/data-planes'
 import { services as externalServicesModule } from '@/app/external-services'
 import { services as gatewaysModule } from '@/app/gateways'
 import { services as legacyDataplanes } from '@/app/legacy-data-planes'
-import { services as meshIdentity } from '@/app/mesh-identity'
-import { services as meshTrust } from '@/app/mesh-trust'
+import { services as meshIdentities } from '@/app/mesh-identities'
+import { services as meshTrusts } from '@/app/mesh-trusts'
 import { services as policies } from '@/app/policies'
 import { services as resources } from '@/app/resources'
 import { services as rules } from '@/app/rules'
@@ -21,6 +21,7 @@ import { services as servicesModule } from '@/app/services'
 import { services as workloads } from '@/app/workloads'
 import type { ServiceDefinition } from '@kumahq/container'
 import type { RouteRecordRaw } from 'vue-router'
+
 
 type Token = ReturnType<typeof token>
 
@@ -98,8 +99,8 @@ export const services = (app: Record<string, Token>): ServiceDefinition[] => {
     ...rules(mesh),
     ...workloads(mesh),
     ...resources(mesh),
-    ...meshIdentity(mesh),
-    ...meshTrust(mesh),
+    ...meshIdentities(mesh),
+    ...meshTrusts(mesh),
   ]
 }
 export const TOKENS = $
