+++ packages/kuma-gui/src/app/resources/views/MeshIdentitySummaryView.vue
@@ -74,7 +74,7 @@
 
 <script lang="ts" setup>
 import { YAML } from '@/app/application'
-import { sources } from '@/app/mesh-identity/sources'
+import { sources } from '@/app/mesh-identities/sources'
 
 const props = defineProps<{
   routeName: string
